Changes to visualize only a portion of the simulated map that is spanned by laser scans.
git-svn-id: https://svn.cc.gatech.edu/borg/OccupancyGrid/trunk@18800 898a188c-9671-0410-8e00-e3fd810bbb7f

+++ src/convert_stanford_log.cpp
@@ -48,7 +48,7 @@ int main(int argc, const char *argv[])
           double angle = min_angle + i * (max_angle - min_angle) / (count - 1);
           range.at<double>(0, i) = r;
           angles.at<double>(0, i) = angle;
-          reflectance.at<double>(0, i) = 1;
+          reflectance.at<uint8_t>(0, i) = 1;
         }
         laser_pose_all.push_back(current_pose);
         laser_range_all.push_back(range);
